refactor(calc): Buttonという名前がmodel層に含まれるのは適切じゃないので改名した。
