--- app/src/main/java/com/example/yosuke/calculator/model/buttons/CalcButton.kt
@@ -1,17 +1,0 @@
-package com.example.yosuke.calculator.model.buttons
-
-interface CalcButton
-
-enum class Operators : CalcButton {
-    PLUS,
-    MINUS,
-    TIMES,
-    DIVIDE,
-    EQUAL;
-}
-
-enum class Specials : CalcButton {
-    CLEAR,
-    SWITCH,
-    PERCENT;
-}
