+++ app/src/main/java/com/example/yosuke/calculator/model/elements/Controller.kt
@@ -1,0 +1,17 @@
+package com.example.yosuke.calculator.model.elements
+
+interface Controller
+
+enum class Operators : Controller {
+    PLUS,
+    MINUS,
+    TIMES,
+    DIVIDE,
+    EQUAL;
+}
+
+enum class Specials : Controller {
+    CLEAR,
+    SWITCH,
+    PERCENT;
+}
