+++ app/src/main/java/com/example/yosuke/calculator/model/usecase/CalcUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.yosuke.calculator.model.usecase
+
+import com.example.yosuke.calculator.model.elements.Operators
+
+interface CalcUseCase {
+    fun calc(x: Long, operator: Operators, y: Long)
+}
