+++ app/src/main/java/com/example/yosuke/calculator/model/usecase/CalcUsecaseImpl.kt
@@ -1,0 +1,9 @@
+package com.example.yosuke.calculator.model.usecase
+
+import com.example.yosuke.calculator.model.elements.Operators
+
+class CalcUsecaseImpl : CalcUseCase {
+    override fun calc(x: Long, operator: Operators, y: Long) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+}
